Use consistent naming in documentation and show dependencies
I haven't used ts-rest yet but when evaluating whether it would be a
good fit I was confused by where `initServer` and `api` were coming
from.

This commit renames `api` to `contract` for next.js to be consistent
with the other library usages and adds a couple comments on where it
comes from.

It also adds the import of initServer in the quick start so it's clear
that's something ts-rest provides.